垃圾分类
Image classification is used to complete the garbage classification project, which is a classic application of deep learning. Its input is a picture, and after some processing, it enters a deep learning model, which will return the garbage category in the image. Here we consider four categories: dry, wet, hazardous or recyclable. Keras package and tensorflow back end are used to build the model.